1. 필수 요구사항 MVC 모델 코드 점검 2. 선택 요구사항 기능 구현

+++ src/main/java/org/example/View/ProductView.java
@@ -19,7 +19,7 @@ public class ProductView {
         int number = 1;
         double price = 0;
         for (Product product : productArrayList){
-            price = product.getPrice()/1000;
+            price = (double)product.getPrice()/1000;
             System.out.println(number++ +". " + product.getName() + "\t| W " + price + "\t| " + product.getDescription());
         }
     }
